merge multiple motifs

merge multiple motifs by calculate mean of each position

mergeMotifs(..., bgNoise = NA)

Arguments

...

pcm or pfm objects

bgNoise

if it is not NA, test will using a background by Dirichlet(1)-distributed random frequencies with weight bg.noise. The value of bgNoise should be a number in the range of 0 to 1, eg. 0.05

Value

a pfm object

Author

Jianhong Ou

Examples

    pcms<-readPCM(file.path(find.package("motifStack"), "extdata"),"pcm$")
    mergeMotifs(pcms)
#> An object of class "pfm"
#> Slot "mat":
#>        [,1]       [,2]       [,3]      [,4]      [,5]        [,6]       [,7]
#> A 0.1481481 0.11111111 0.45833333 0.1602769 0.3604588 0.152188373 0.93740032
#> C 0.1481481 0.03703704 0.19391026 0.3434505 0.1407557 0.123985633 0.04837852
#> G 0.1481481 0.00000000 0.26335470 0.1512247 0.4816464 0.001064254 0.00000000
#> T 0.5555556 0.85185185 0.08440171 0.3450479 0.0171390 0.722761740 0.01422116
#>          [,8]         [,9]       [,10]        [,11]     [,12]
#> A 0.984848485 0.9997341839 0.000000000 0.9917487357 0.5227464
#> C 0.012626263 0.0002658161 0.913343966 0.0011977642 0.1453243
#> G 0.000000000 0.0000000000 0.005582137 0.0001330849 0.1356361
#> T 0.002525253 0.0000000000 0.081073897 0.0069204152 0.1962932
#> 
#> Slot "name":
#> [1] "FoxP_SOLEXA;bin_SOLEXA;fd64A_SOLEXA;fkh_NAR;foxo_SOLEXA;slp1_SOLEXA;slp2_SOLEXA"
#> 
#> Slot "alphabet":
#> [1] "DNA"
#> 
#> Slot "color":
#>         A         C         G         T 
#> "#00811B" "#2000C7" "#FFB32C" "#D00001" 
#> 
#> Slot "background":
#>    A    C    G    T 
#> 0.25 0.25 0.25 0.25 
#> 
#> Slot "tags":
#> list()
#> 
#> Slot "markers":
#> list()
#>